Markdown fixes (#736)
* Fixed Number System readme.md

Fixed markdown formatting for clarity and readability.
Modified charts for better readability.
Edited some explanations to match chart changes and for clarity.
Fixed table of contents and section title links so they work.
Fixed file name for lua example so the link to it works.
Fixed minor spelling errors.

* Worked on formatting

I went through all the data structures pages and added or fixed a table of contents if the page was more than about a page and a half. There were some pages that didn't have use case examples for the data structures, so I added those as well.